Admin/requirements.txt
streamlit
pypdf
langchain
langchain_community
faiss_cpu
boto3
python-dotenv